added readme added error handling in mediatr handlers added unit tests

+++ DemoUsers.Server/Users/Features/DeleteUser/DeleteUserCommandHandler.cs
@@ -8,10 +8,10 @@ namespace DemoUsers.Server.Users.Features.DeleteUser
 
     internal class DeleteUserCommandHandler : IRequestHandler<DeleteUserCommand, bool>
     {
-        readonly ILogger _logger;
+        readonly ILogger<DeleteUserCommandHandler> _logger;
         readonly IUsersRepository _usersRepository;
 
-        public DeleteUserCommandHandler(ILogger<CreateUserCommandHandler> logger, IUsersRepository usersRepository)
+        public DeleteUserCommandHandler(ILogger<DeleteUserCommandHandler> logger, IUsersRepository usersRepository)
         {
             _logger = logger;
             _usersRepository = usersRepository;
@@ -19,7 +19,30 @@ namespace DemoUsers.Server.Users.Features.DeleteUser
 
         public async Task<bool> Handle(DeleteUserCommand request, CancellationToken cancellationToken)
         {
-            return await _usersRepository.DeleteUserAsync(request.Id, cancellationToken);
+            _logger.LogInformation($"[{nameof(DeleteUserCommandHandler)}] deleting user: {request.Id}");
+            try
+            {
+                var user = await _usersRepository.GetUserAsync(request.Id, cancellationToken);
+                if (user == null)
+                {
+                    _logger.LogWarning($"[{nameof(DeleteUserCommandHandler)}] user not found: {request.Id}");
+                    return false;
+                }
+
+                var result = await _usersRepository.DeleteUserAsync(request.Id, cancellationToken);
+                if (result)
+                    _logger.LogInformation($"[{nameof(DeleteUserCommandHandler)}] user deleted: {request.Id}");
+                else
+                    _logger.LogWarning($"[{nameof(DeleteUserCommandHandler)}] user not deleted: {request.Id}");
+
+                return result;
+            }
+            catch (Exception ex)
+            {
+                _logger.LogError(ex, $"[{nameof(DeleteUserCommandHandler)}] Error occurred while getting user: {request.Id}");
+                throw;
+            }
+            
         }
     }
 }
